+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Draft
+
+### :beetle: Fixes
+
+* Check for both OS and CPU hardware for updates (e.g. differentiate between Intel and Apple Silicon)
+* Include link to release notes in update window
+* Much improved dialog for available updates
+
 ## 1.7.0
 
 ### What's Changed